Changes that exclude namespaces that are optional (i.e. explicitly... ... declared as accepting "jsUndefined" or implicitly by virtue of declaring a default value) from including in the merged filter spec digraph model. This patch additionally checks and ensures that each filter passed into the discriminator constructor uses a unique operation ID (a longstanding problem).
Note that this patch is _better_ than previous version insofar as
it will not incorrectly consider an undefined namespace as a significant
feature of request input wrt routing. However, the manner in which this
exlusion is made in the patch is not correct. We cannot simply exclude
conditional namespaces from inclusion in the merged filter spec digraph
model because we need to ensure discriminator does not incorrectly conclude
that the same namespace declared by another filter is not assumed to be
a reasonable candidate for discriminating messages by virtue of the fact
that there's apparently no conflict in the merged digraph model.

Also, there's similar but a bit more difficult-to-solve issue that involves
keeping track of the mutually-exclusive sets of types a specific filter/namespace
defines. We must ensure that any filter/namespace that declares itself as accepting
more than a single type be treated as a mutually exclusive set. If another filter
defines its own type set for the same namespace then these sets must be disjoint
or the namespace cannot be discriminated.